pyproject.toml
[tool.poetry]
name = "LickingDogAPI"
version = "1.0.0"
description = "舔狗API"
authors = ["llxlr <i@xhlr.top>"]
homepage = "https://api.white-album.top/"
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8,<4.0"

aiofiles = ">=0.5.0"
aiohttp = "^3.7.4.post0"
aioredis = "^1.3.1"
amzqr = "^0.0.1"
baidu-aip = "^2.2.18.0"
beautifulsoup4 = "^4.9.3"
bilibili_api = "^7.0.1"
celery = "^5.1.2"
cffi = "^1.14.5"
click = "^7.1.2"
cssutils = "^1.0.2"
email-validator = "^1.1.2"
eliot = "^1.13.0"
eliot-tree = "^21.0.0"
fake-useragent = "^0.1.11"
faust = "^1.10.4"
gunicorn = "^20.1.0"
Jinja2 = "^2.11.3"
loguru = "^0.5.3"
lxml = "^4.6.2"
numpy = "^1.20.3"
# opencv-python = "^4.4.0.46"
orjson = "^3.5.3"
pandas = "^1.1.5"
passlib = {version = "^1.7.4", extras = ["bcrypt"]}
Pillow = "^8.3.1"
pixivpy = "^3.6.0"
#pixivpy-async = "^1.2.12"
# playwright = "^0.171.1"
pydantic = "^1.8.2"
PyMySQL = "^0.9.3"
# pysnooper = "^0.4.2"
python-dotenv = "^0.18.0"
python-jose = {version = "^3.3.0", extras = ["cryptography"]}
python-multipart = "^0.0.5"
requests = "^2.25.1"
slowapi = "^0.1.4"
SQLAlchemy = "^1.4.21"
starlette = ">=0.14.2"
twint = "^2.1.20"
uvicorn = {version = "^0.14.0", extras = ["standard"]}
yagmail = "^0.14.256"
pyppeteer = "^0.2.6"
fastapi = "^0.70.0"
selenium = "<4.0.0"
strawberry-graphql = "^0.83.4"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"
poethepoet = "^0.10.0"
pytest-cov = "^2.12.1"
pytest-xdist = "^2.3.0"

[[tool.poetry.source]]
name = "bfsu"
url = "https://mirrors.bfsu.edu.cn/pypi/web/simple"
default = true

[build-system]
requires = ["poetry-core>=1.0.0", "setuptools"]
build-backend = "poetry.core.masonry.api"